requirements.txt
pytest
pytest-playwright
playwright
pytest-xdist
python-dotenv
pyyaml
allure-pytest